+++ crates/astria-composer/src/executor/mod.rs
@@ -87,7 +87,7 @@ type StdError = dyn std::error::Error;
 /// an `Unsigned`, then signs them with the sequencer key and submits to the sequencer.
 /// Its `status` field indicates that connection to the sequencer node has been established.
 #[derive(Debug)]
-pub(super) struct Executor {
+pub(crate) struct Executor {
     // The status of this executor
     status: watch::Sender<Status>,
     // Channel for receiving `SequenceAction`s to be bundled.
@@ -135,8 +135,8 @@ impl Drop for Executor {
 }
 
 #[derive(Debug)]
-pub(super) struct Status {
-    is_connected: bool,
+pub(crate) struct Status {
+    pub(crate) is_connected: bool,
 }
 
 impl Status {
@@ -146,13 +146,13 @@ impl Status {
         }
     }
 
-    pub(super) fn is_connected(&self) -> bool {
+    pub(crate) fn is_connected(&self) -> bool {
         self.is_connected
     }
 }
 
 impl Executor {
-    pub(super) fn new(
+    pub(crate) fn new(
         sequencer_url: &str,
         private_key: &SecretString,
         block_time: u64,
@@ -188,7 +188,7 @@ impl Executor {
     }
 
     /// Return a reader to the status reporting channel
-    pub(super) fn subscribe(&self) -> watch::Receiver<Status> {
+    pub(crate) fn subscribe(&self) -> watch::Receiver<Status> {
         self.status.subscribe()
     }
 
@@ -212,7 +212,7 @@ impl Executor {
     /// # Errors
     /// An error is returned if connecting to the sequencer fails.
     #[instrument(skip_all, fields(address = %self.address))]
-    pub(super) async fn run_until_stopped(mut self) -> eyre::Result<()> {
+    pub(crate) async fn run_until_stopped(mut self) -> eyre::Result<()> {
         let mut submission_fut: Fuse<Instrumented<SubmitFut>> = Fuse::terminated();
         let mut nonce = get_latest_nonce(self.sequencer_client.clone(), self.address)
             .await
